working tabs footer and better look overall
