+++ app/src/components/Footer.svelte
@@ -1,0 +1,9 @@
+<footer class="bg-black text-white p-3 fixed bottom-0 text-xs">
+    <div class="container mx-auto">
+    <p class="text-center">&copy; {new Date().getFullYear()} {club_name}</p>
+    </div>
+</footer>
+
+<script>
+    import { club_name  } from "../config";
+</script>
