+++ app/src/config.ts
@@ -1,1 +1,2 @@
 export const base_api: string = import.meta.env.API_URL || 'http://localhost:3000/';
+export const club_name: string = import.meta.env.CLUB_NAME || 'Koło Naukowe Data Science';
