+++ app/src/routes/+page.svelte
@@ -3,19 +3,15 @@
     <div class="absolute top-0 left-0 w-full h-full">
             <h1 class="text-3xl text-red-600 my-3 text-center">Dashboard</h1>
             <div class="text-4xl text-yellow-600 my-3 text-center">Your models:</div>
-        <Models position="absolute"/>
+        <Models/>
     </div>
 </div>
 <Background position="relative"/>
-
-<style>
-    :global(body) {
-        background-color: #FEE2E2;
-    }
-</style>
+<Footer/>
 
 <script lang="ts">
-    import Navbar from '../Navbar.svelte';
-    import Models from '../Models.svelte';
-    import Background from '../Background.svelte';
+    import Navbar from '../components/Navbar.svelte';
+    import Models from '../components/Models.svelte';
+    import Background from '../components/Background.svelte';
+    import Footer from '../components/Footer.svelte';
 </script>
