+++ app/src/routes/dashboard/+page.svelte
@@ -1,0 +1,5 @@
+<MainPage/>
+
+<script lang="ts">
+    import MainPage from '../+page.svelte'
+</script>
