+++ app/src/routes/datasets/+page.svelte
@@ -1,7 +1,7 @@
 <Navbar/>
 <div class="relative">
     <div class="absolute top-0 left-0 w-full h-full">
-        <Models/>
+        
     </div>
 </div>
 <Background position="relative"/>
@@ -9,7 +9,6 @@
 
 <script lang="ts">
     import Navbar from '../../components/Navbar.svelte';
-    import Models from '../../components/Models.svelte';
     import Background from '../../components/Background.svelte';
     import Footer from '../../components/Footer.svelte';
 </script>
