+++ app/src/routes/model metrics/+page.svelte
@@ -1,7 +1,7 @@
 <Navbar/>
 <div class="relative">
     <div class="absolute top-0 left-0 w-full h-full">
-        <Models/>
+
     </div>
 </div>
 <Background position="relative"/>
